Add public top level types to cache in java-imports-scan-file
Add new function java-imports-list-top-level and use it in
java-imports-scan-file to add public top level types to list
of classes to add to cache.

Useful when adding new types or working with a new project,
save having to specify the package on first import.
Only works if java-imports-scan-file runs after type is declared
in file so for new files it might be worth adding
(java-imports-scan-file) to define-auto-insert for example.

No kotlin-mode support yet, haven't figured out a good way
to do a "negative" match as "public" modifier is the default
in kotlin. Also the primary constructor syntax is tricky to include
and the regex is already rather horrible.